Refactored nested replacer to not need logs
